Fix performance regression introduced by #7898
As per
https://docs.rs/tokio-threadpool/0.1.15/tokio_threadpool/fn.blocking.html
when you use a blocking block in a task, it blocks the _entire_ task,
not just the bit which you're marking blocking. This means that when
we're in one giant join_all calling read_dir on a bunch of directories,
we end up serializing them to only read one directory at a time.

This change introduces a new logging::Executor (a slightly weird place
for it, but it depends on the logging stuf, so...) which mirrors the
tokio::Runtime and the futures::sync::oneshot::spawn APIs to
conveniently allow blocking Futures to be spawned in their own tasks, so
that they don't hinder parallelism within a task.