Export main class in `jvm_binary` targets
Previously, the export from `export-fastpass` didn't include the name of the
main class for `jvm_binary` targets. For targets with multiple `main`s, or
targets whose `main` is defined in a dependency, couldn't easily be run by
Bloop (the main class would have to be manually specified).

This commit adds this information for `jvm_binary` targets.